app/requirements/production.txt
django
django-crispy-forms
django-allauth
markdown
django-filter
django-pwa
gunicorn
tzdata
psycopg2-binary